API/requirements.txt
python-dotenv>=1.0.0
flask>=2.3.0
flask-socketio>=5.3.0
supabase>=2.0.0
requests>=2.32.5
google-generativeai>=0.8.0